feat: update amount type from int to long across loot tracking components

+++ AlbionDataAvalonia/Loot/Models/LootModels.cs
@@ -26,7 +26,7 @@ public sealed record LootRecord(
     string ItemUniqueName,
     string ItemName,
     int? Quality,
-    int Amount,
+    long Amount,
     long? EstimatedMarketValue,
     long? TotalEstimatedMarketValue);
